update api user

+++ restfullApi/api/create_user.php
@@ -6,15 +6,12 @@
   header("Access-Control-Max-Age: 3600");
   header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
 
-  // files needed to connect to database
   include_once 'config/database.php';
   include_once 'objects/user.php';
-  
-  // get database connection
+
   $database = new Database();
   $db = $database->getConnection();
-  
-  // instantiate product object
+
   $user = new User($db);
 
   // get posted data
@@ -25,24 +22,33 @@
   $user->lastname = $data->lastname;
   $user->email = $data->email;
   $user->password = $data->password;
-  
-  // create the user
-  if($user->create()){
-  
-    // set response code
-    http_response_code(200);
+  $user->gender = $data->gender;
+  $user->birthday = $data->birthday;
 
-    // display message: user was created
-    echo json_encode(array("message" => "User was created."));
+  if (!($user->emailExists())){
+      // create the user
+    if($user->create()){
+    
+      // set response code
+      http_response_code(200);
+
+      // display message: user was created
+      echo json_encode(array("message" => "User was created."));
+    }
+
+    // message if unable to create user
+    else{
+
+      // set response code
+      http_response_code(400);
+
+      // display message: unable to create user
+      echo json_encode(array("message" => "Unable to create user."));
+    }
   }
-
-  // message if unable to create user
-  else{
-
-    // set response code
-    http_response_code(400);
-
-    // display message: unable to create user
-    echo json_encode(array("message" => "Unable to create user."));
+  else {
+    echo json_encode(array("message" => "Email exists"));
   }
+  
+  
 ?>
